存在しないタグの処理を追加

+++ app/models/remove_tags_api.rb
@@ -10,9 +10,11 @@ class RemoveTagsApi < ApiBase
   # FIXME: tag1の長さを検証
 
   def execute(user_id)
-    tag = Tag.get(self.tag1)
-    tagging = Tagging.find_by_user_id_and_article_id_and_tag_id(user_id, self.article_id, tag.id)
-    tagging.destroy
+    tag = Tag.get(self.tag1, :create => false)
+    if tag
+      tagging = Tagging.find_by_user_id_and_article_id_and_tag_id(user_id, self.article_id, tag.id)
+      tagging.destroy
+    end
 
     return {
       :success => true,
